SortingVisualizer_0.20
Generation functions and the definition of certain data generating/inserting functions have been defined. the user is able to choose between the two options in a very similar manner to how they choose the algorithm Data Type in algorithmDataType(). the actual generation and transfer of data has not been implemented yet, but will be in SV_0.30/0.40