Sorting_Visualizer_0.301
some modifications have been applied to the program to allow for inputCharacter() to return an address to wherever its called, removing the need to free() every time that function is called. A double LL Structure is being implemented to store data, it will also allow for things to be rearranged more easily. data_Type will need to be altered, since I'm receiving an error when comparing things (see line 74 in doubleLLFunctions.c) and I'm unsure why that error is occuring. It may be likely that could be a result of having data_Type being a union structure