SortingVisualizer_0.302
The find() and remove() functions of the DoublyLL have been implemented, although testing of such functions is required, all other DLL functions must be completed first. the method of storing data in the DATA_NODE(s) still needs to be worked on, since they now handle only char*, or an array of strings.