Sorting_Visualizer_0.303
all functions of doubleLLFunctions have been implemented, although testing it is neccessary, it would be wiser to complete data collection and storage prior to testing. Version_0.305 should be able to store and manage a single data type, which makes some segments of code irrelevant, but other data types will be implemented as the program is being constructed.