Sorting_Visualizer_0.305
Data collection has been implemented, with a few simple checks to ensure data integrity. Due to the generalization of the doubly list data structure, where nodes have a generic pointer to data, a pointer of any data type can be passed during the insertion process into the doubly list. This comes with a dangerous, but powerful outcome: a single list can store multiple different types of data at once. This could cause some complications with the implementation of the sorting algorithms, but a definition of the data type a doubly list  is allowed to store could be a very easy fix. Its structure definition just needs to be modified slightly.  All necessary functions - with the exception of the sorting algorithms themselves - have been implemented. Testing is yet to be done, but  is expected to be at least able to compile by version 0.400.