SortingVisualizer_0.450
All errors pertaining to the doubly list have been resolved. Inputting, generating, storing, and displaying data items works for all three data types, although in rare occasions a string of gibberish is passed to the doubly list during the generation of asterisk bars, I'm unsure as to why and there seems to be no pattern for when it occurs. version 0.500 should present at least 1 sorting algorithm, but it may not display a step by step visualization for the time being.