SortingVisualizer_0.600
Some minor changes were made to .vscode\ files to allow for an easier way of building the program. I should look into making this more user-friendly so anyone can load up my code to modify & build it with ease, we'll see.

Lots of changes have been made to all 4 sorting algorithms, the key aspect being how data is being moved. previously the nodes themselves weren't moving, but the data they held was being swapped. this has been resolved in a way so that the actual nodes are being rearranged within the DLL.

Additionally, a simple addition of Sleep() & system("cls") has made the program actually show what's being moved in every step of the sorting process, it just need to be more visually appealing and some form of highlighting to which item was moved should be used.