SortingVisualizer_0.700
ANSI escape codes have been introduced! I've developed a simple library that allows me to use a few simple ANSI codes to format the terminal output more neatly, they're simply inserted before & after a string of text inside a printf() statement. Ideally by version 1.000, a complete overhaul of the program's TUI will be completed, but there is still a lot of work to be done with the current formatting. a method to highlight what's currently being compared and what was just recently moved needs to be added somehow.